+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - `MessageLifecycle::failed` no longer has an `$isRetrying` argument, instead
   `MessageLifecycyle::retrying` will be called instead.
+- The `Message` interface no longer has a `getName` method.
+  Instead the *name* of a message is its fully qualified class name. Should the
+  old behavior still be desired, implement `PMG\Queue\NamedMessage` instead.
 
 ### Fixed
 
@@ -20,6 +23,13 @@ n/a
   a message fails and is retrying.
 - `PMG\Queue\Lifecycle\DelegatingLifecycle` has a new named constructor:
   `fromIterable`. This uses PHP 7.1's `iterable` pseudo type.
+- A `PMG\Queue\NamedMessage` interface to enable the old behavior of
+  `Message::getName`.
+
+### Removed
+
+- The `PMG\Queue\MessageTrait` has been removed. The behavior it provided (using
+  the fully qualified class name as the message name) is now the default.
 
 ### Deprecations